+++ test/Myriad.IntegrationPluginTests/FieldsLensDus.fs
@@ -35,7 +35,9 @@ module RecordWithEmptyWrapperNameLenses =
 
     let one_empty_wrapper_name =
         (fun (x: RecordWithEmptyWrapperName) -> x.one_empty_wrapper_name),
-        (fun (x: RecordWithEmptyWrapperName) (value: int) -> { x with one_empty_wrapper_name = value })
+        (fun (x: RecordWithEmptyWrapperName) (value: int) ->
+            { x with
+                one_empty_wrapper_name = value })
 namespace rec TestLens
 
 module RecordWithWrappedLensViaTypedefofLenses =